[#1415] Modify Test
