+++ agent-it/src/test/java/com/navercorp/pinpoint/plugin/mongodb/MongoDB_3_7_x.java
@@ -16,13 +16,15 @@
 
 package com.navercorp.pinpoint.plugin.mongodb;
 
-import com.mongodb.MongoClientOptions;
+import com.mongodb.MongoClientSettings;
+import com.mongodb.MongoDriverInformation;
 import com.mongodb.ReadPreference;
 import com.mongodb.WriteConcern;
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.result.DeleteResult;
-import com.mongodb.connection.Cluster;
 import com.navercorp.pinpoint.bootstrap.plugin.test.ExpectedAnnotation;
 import com.navercorp.pinpoint.bootstrap.plugin.test.PluginTestVerifier;
 import com.navercorp.pinpoint.bootstrap.plugin.test.PluginTestVerifierHolder;
@@ -32,14 +34,6 @@ import com.navercorp.pinpoint.common.util.JsonParser;
 import com.navercorp.pinpoint.common.util.NormalizedJson;
 import com.navercorp.pinpoint.test.plugin.Dependency;
 import com.navercorp.pinpoint.test.plugin.PinpointPluginTestSuite;
-import de.flapdoodle.embed.mongo.MongodExecutable;
-import de.flapdoodle.embed.mongo.MongodProcess;
-import de.flapdoodle.embed.mongo.MongodStarter;
-import de.flapdoodle.embed.mongo.config.IMongodConfig;
-import de.flapdoodle.embed.mongo.config.MongodConfigBuilder;
-import de.flapdoodle.embed.mongo.config.Net;
-import de.flapdoodle.embed.mongo.distribution.Version;
-import de.flapdoodle.embed.process.runtime.Network;
 import org.bson.Document;
 import org.bson.conversions.Bson;
 import org.junit.AfterClass;
@@ -48,9 +42,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.lang.reflect.Constructor;
 import java.lang.reflect.Method;
-import java.util.List;
 
 import static com.navercorp.pinpoint.bootstrap.plugin.test.Expectations.event;
 import static com.navercorp.pinpoint.bootstrap.plugin.test.Expectations.json;
@@ -60,13 +52,13 @@ import static com.navercorp.pinpoint.bootstrap.plugin.test.Expectations.json;
  */
 @RunWith(PinpointPluginTestSuite.class)
 @Dependency({
-        "org.mongodb:mongodb-driver:[3.0.0,3.6.max]",
-        "org.mongodb:bson:3.6.4",
+        "org.mongodb:mongodb-driver:[3.7.0,3.8.max]",
+        "org.mongodb:bson:3.7.0",
         "de.flapdoodle.embed:de.flapdoodle.embed.mongo:1.50.5"
         })
-public class MongoDB_3_0_x extends MongoDBBase {
+public class MongoDB_3_7_x extends MongoDBBase {
 
-    private static final String MONGODB_VERSION = "3_0_x";
+    private static final String MONGODB_VERSION = "3_7_x";
 
     // for normalized json
     protected static final JsonParser JSON_PARSER = new DefaultJsonParser();
@@ -87,11 +79,13 @@ public class MongoDB_3_0_x extends MongoDBBase {
 
         PluginTestVerifier verifier = PluginTestVerifierHolder.getInstance();
 
-        com.mongodb.MongoClient mongoClient = new com.mongodb.MongoClient( "localhost" , 27018);
-        verifier.printCache();
-        Class<?> clusterClass = Class.forName("com.mongodb.Mongo");
+        //create DB
+        MongoClient mongoClient = MongoClients.create( "mongodb://localhost:27018" );
 
-        Constructor create = clusterClass.getDeclaredConstructor(Cluster.class, MongoClientOptions.class, List.class);
+        verifier.printCache();
+        Class<?> CreateDBClass = Class.forName("com.mongodb.client.MongoClients");
+
+        Method create = CreateDBClass.getDeclaredMethod("create", MongoClientSettings.class, MongoDriverInformation.class);
         verifier.verifyTrace(event(MONGODB, create, null, MONGODB_ADDRESS, null));
 
         database = mongoClient.getDatabase("myMongoDbFake").withReadPreference(ReadPreference.secondaryPreferred()).withWriteConcern(WriteConcern.MAJORITY);
@@ -101,14 +95,15 @@ public class MongoDB_3_0_x extends MongoDBBase {
         Document doc = new Document("name", "Roy").append("company", "Naver");
         collection.insertOne(doc);
 
-        Class<?> mongoDatabaseImpl = Class.forName("com.mongodb.MongoCollectionImpl");
+        Class<?> mongoDatabaseImpl = Class.forName("com.mongodb.client.internal.MongoCollectionImpl");
         Method insertOne = mongoDatabaseImpl.getDeclaredMethod("insertOne", Object.class);
         String nosql = doc.toJson();
         NormalizedJson normalizedJson = JSON_PARSER.normalizeJson(nosql);
 
         verifier.verifyTrace(event(MONGO_EXECUTE_QUERY, insertOne, null, MONGODB_ADDRESS, null
-                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers using MAJORITY")
+                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers with MAJORITY")
                 , json(normalizedJson.getNormalizedJson(), normalizedJson.getParseParameter())));
+
 
         //insert Data
         MongoCollection<Document> collection2 = database.getCollection("customers2").withWriteConcern(WriteConcern.ACKNOWLEDGED);
@@ -120,8 +115,9 @@ public class MongoDB_3_0_x extends MongoDBBase {
         NormalizedJson normalizedJson2 = JSON_PARSER.normalizeJson(nosql2);
 
         verifier.verifyTrace(event(MONGO_EXECUTE_QUERY, insertOne, null, MONGODB_ADDRESS, null
-                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers2 using ACKNOWLEDGED")
+                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers2 with ACKNOWLEDGED")
                 , json(normalizedJson2.getNormalizedJson(), normalizedJson2.getParseParameter())));
+
 
         //update Data
         Document doc3 = new Document("$set", new Document("name", "Roy3"));
@@ -132,7 +128,7 @@ public class MongoDB_3_0_x extends MongoDBBase {
 
         Method updateOne = mongoDatabaseImpl.getDeclaredMethod("updateOne", Bson.class , Bson.class);
         verifier.verifyTrace(event(MONGO_EXECUTE_QUERY, updateOne, null, MONGODB_ADDRESS, null
-                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers using MAJORITY")
+                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers with MAJORITY")
                 , json(normalizedJson3.getNormalizedJson(), normalizedJson3.getParseParameter())));
 
         //read data
@@ -140,8 +136,8 @@ public class MongoDB_3_0_x extends MongoDBBase {
         Method find = mongoDatabaseImpl.getDeclaredMethod("find");
 
         verifier.verifyTrace(event(MONGO_EXECUTE_QUERY, find, null, MONGODB_ADDRESS, null
-                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers using SECONDARYPREFERRED")
-        ));
+                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers with SECONDARYPREFERRED")
+                ));
 
         int resultCount=0;
         try {
@@ -154,19 +150,19 @@ public class MongoDB_3_0_x extends MongoDBBase {
         }
         Assert.assertEquals(1, resultCount);
 
-        //delete
+        //delete data
         Document doc4 = new Document("name", "Roy3");
         NormalizedJson normalizedJson4 = JSON_PARSER.normalizeJson(doc4.toJson());
 
         DeleteResult deleteResult = collection.deleteMany(doc4);
         Method deleteMany = mongoDatabaseImpl.getDeclaredMethod("deleteMany", Bson.class);
         verifier.verifyTrace(event(MONGO_EXECUTE_QUERY, deleteMany, null, MONGODB_ADDRESS, null
-                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers using MAJORITY")
+                , new ExpectedAnnotation(AnnotationKey.MONGO_COLLECTIONINFO.getName(), "customers with MAJORITY")
                 , json(normalizedJson4.getNormalizedJson(), normalizedJson4.getParseParameter())));
 
         Assert.assertEquals(1, deleteResult.getDeletedCount());
 
-
+        //close connection
         mongoClient.close();
         stopDB();
     }
